Align Header text to baseline
